estrela-server/requirements.txt
astroid==1.6.4
autopep8==1.3.5
Django==2.0.5
django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-queryfields==1.0.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
Pillow==5.2.0
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==1.9.1
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
wrapt==1.10.11
